chore: 인스타 랜딩 링크에서 utm 파라미터·open 서브도메인 제거 (#517)
* chore: 인스타 랜딩 유입 소스를 UA 로 판정해 utm 파라미터 제거

인스타 바이오에 `https://piki.day/open` 만 걸 수 있도록, `?utm_source=instagram` 없이도
인앱 브라우저 UA(`isInstagramBrowser`)로 유입 소스를 판정한다.

- 바이오 탭은 인스타 인앱 브라우저로 열리므로 랜딩 도달 시점에 UA 토큰이 남아 있다
- 쿼리로 들어온 `utm_source` 가 있으면 그쪽이 우선 — 다른 채널 확장 시 그대로 사용
- GA4 `landing_view` 의 `source`, Play 스토어 `referrer` 는 동일하게 유지된다

* chore: 미사용 open.piki.day 랜딩 서브도메인 제거

인스타 링크를 `piki.day/open` 으로 직접 걸기로 해서 서브도메인 경로를 걷어낸다.

- `proxy.ts` 의 `open.*` 루트 rewrite 제거
- `isLandingHost` 삭제 — 앱의 associatedDomains 에 등록된 적 없는 호스트였다
- `landingHost.ts` → `serviceHost.ts` 로 이름 변경. 남은 `toServiceHost` 는
  Vercel 프리뷰 등 미등록 호스트를 프로덕션으로 정규화하는 역할만 한다

* fix: Host 헤더를 엄격히 파싱해 userinfo 오리진 조작 차단
