+++ apps/web/src/app/open/page.tsx
@@ -2,9 +2,9 @@ import type { Metadata } from 'next';
 import { headers } from 'next/headers';
 import { redirect } from 'next/navigation';
 
-import { DEFAULT_LANDING_TARGET } from '@/consts/appLink';
+import { DEFAULT_LANDING_TARGET, INSTAGRAM_SOURCE } from '@/consts/appLink';
 import isSafeInternalPath from '@/utils/isSafeInternalPath';
-import { toServiceHost } from '@/utils/landingHost';
+import { toServiceHost } from '@/utils/serviceHost';
 import { isWebview } from '@/utils/webBridge';
 
 import OpenLanding from './_components/OpenLanding';
@@ -39,12 +39,20 @@ async function OpenPage({ searchParams }: OpenPageProps) {
     redirect(`${serviceOrigin}${target}`);
   }
 
+  /**
+   * 유입 소스 — 인스타 바이오에 `?utm_source=instagram` 없이 깔끔한 링크를 걸기 위해
+   * 인앱 브라우저 UA 로도 판정한다. 바이오 탭은 인스타 인앱 브라우저로 열리므로
+   * 이 랜딩에 도달하는 시점엔 아직 UA 에 `Instagram` 토큰이 남아 있다.
+   * 쿼리로 들어온 값이 있으면 그쪽이 우선 — 다른 채널을 붙일 때 그대로 쓴다.
+   */
+  const source = utmSource ?? (landingEnv.isInstagramBrowser ? INSTAGRAM_SOURCE : null);
+
   return (
     <OpenLanding
       landingEnv={landingEnv}
       target={target}
       serviceOrigin={serviceOrigin}
-      source={utmSource ?? null}
+      source={source}
     />
   );
 }
